+++ HISTORY.md
@@ -1,3 +1,9 @@
+unreleased
+==========
+
+  * deps: qs@1.2.0
+    - Fix parsing array of objects
+
 1.6.1 / 2014-08-06
 ==================